+++ api/aecworks/settings.py
@@ -70,6 +70,7 @@ INSTALLED_APPS = [
     "api.users",
     "api.community",
     "api.images",
+    "api.social",
 ]
 
 MIDDLEWARE = [
