+++ api/community/views/company.py
@@ -5,7 +5,6 @@ from rest_framework import (
     generics,
     serializers,
     permissions,
-    # filters,
 )
