+++ api/users/urls.py
@@ -7,7 +7,6 @@ urlpatterns = [
     path("profiles/", views.ProfileListView.as_view()),
     path("profiles/me/", views.ProfileMeView.as_view()),
     path("profiles/<str:slug>/", views.ProfileDetailView.as_view()),
-    path("profiles/<str:slug>/tweets/", views.ProfileTweetsView.as_view()),
     path("login/<str:provider_name>/", views.OauthLoginView.as_view()),
     path("logout/", views.Logout.as_view()),
     path("login/", views.Login.as_view()),
